chore: enable react-refresh/only-export-components
