+++ client/src/app/pages/vulnerability-list/vulnerability-context.ts
@@ -1,0 +1,25 @@
+import React from "react";
+
+import type { AxiosError } from "axios";
+
+import type { VulnerabilitySummary } from "@app/client";
+import type { ITableControls } from "@app/hooks/table-controls";
+
+export interface IVulnerabilitySearchContext {
+  tableControls: ITableControls<
+    VulnerabilitySummary,
+    "identifier" | "title" | "severity" | "published" | "sboms",
+    "identifier" | "published" | "severity",
+    "" | "base_severity" | "published" | "publishedOnly",
+    string
+  >;
+
+  totalItemCount: number;
+  isFetching: boolean;
+  fetchError: AxiosError | null;
+}
+
+const contextDefaultValue = {} as IVulnerabilitySearchContext;
+
+export const VulnerabilitySearchContext =
+  React.createContext<IVulnerabilitySearchContext>(contextDefaultValue);
