+++ client/src/app/pages/vulnerability-list/vulnerability-list.tsx
@@ -2,7 +2,7 @@ import type React from "react";
 
 import { Content, PageSection } from "@patternfly/react-core";
 
-import { VulnerabilitySearchProvider } from "./vulnerability-context";
+import { VulnerabilitySearchProvider } from "./vulnerability-provider";
 import { VulnerabilityTable } from "./vulnerability-table";
 import { VulnerabilityToolbar } from "./vulnerability-toolbar";
 import { DocumentMetadata } from "@app/components/DocumentMetadata";
