+++ client/src/app/pages/vulnerability-list/vulnerability-provider.tsx
@@ -1,16 +1,12 @@
 import React from "react";
 
-import type { AxiosError } from "axios";
-
 import {
   FILTER_NULL_VALUE,
   FILTER_TEXT_CATEGORY_KEY,
   TablePersistenceKeyPrefixes,
 } from "@app/Constants";
-import type { VulnerabilitySummary } from "@app/client";
 import { FilterType } from "@app/components/FilterToolbar";
 import {
-  type ITableControls,
   getHubRequestParams,
   useTableControlProps,
   useTableControlState,
@@ -18,24 +14,7 @@ import {
 import { useFetchVulnerabilities } from "@app/queries/vulnerabilities";
 import { parseBooleanIfPossible } from "@app/utils/utils";
 
-export interface IVulnerabilitySearchContext {
-  tableControls: ITableControls<
-    VulnerabilitySummary,
-    "identifier" | "title" | "severity" | "published" | "sboms",
-    "identifier" | "published" | "severity",
-    "" | "base_severity" | "published" | "publishedOnly",
-    string
-  >;
-
-  totalItemCount: number;
-  isFetching: boolean;
-  fetchError: AxiosError | null;
-}
-
-const contextDefaultValue = {} as IVulnerabilitySearchContext;
-
-export const VulnerabilitySearchContext =
-  React.createContext<IVulnerabilitySearchContext>(contextDefaultValue);
+import { VulnerabilitySearchContext } from "./vulnerability-context";
 
 interface IVulnerabilityProvider {
   children: React.ReactNode;
